% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_bm25.R
\docType{data}
\name{epi_bm25}
\alias{epi_bm25}
\title{BM25 for the 'epiverse-connect' r-universe}
\format{
An object of class \code{list} of length 2.
}
\usage{
epi_bm25
}
\description{
In the format expected by \code{\link[pkgmatch:pkgmatch_similar_pkgs]{pkgmatch::pkgmatch_similar_pkgs()}}.
}
\keyword{datasets}
