% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_embeddings.R
\docType{data}
\name{epi_embeddings}
\alias{epi_embeddings}
\title{Embeddings for the 'epiverse-connect' r-universe}
\format{
An object of class \code{list} of length 3.
}
\usage{
epi_embeddings
}
\description{
In the format returned by \code{\link[pkgmatch:pkgmatch_embeddings_from_pkgs]{pkgmatch::pkgmatch_embeddings_from_pkgs()}}, and
expected by \code{\link[pkgmatch:pkgmatch_similar_pkgs]{pkgmatch::pkgmatch_similar_pkgs()}}.
}
\keyword{datasets}
