% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs_metadata.R
\docType{data}
\name{pkgs_metadata}
\alias{pkgs_metadata}
\title{Metadata of the packages the 'epiverse-connect' r-universe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 106 rows and 4 columns.
}
\usage{
pkgs_metadata
}
\description{
As returned by \code{\link[epiverse.scraper:get_pkgs_metadata]{epiverse.scraper::get_pkgs_metadata()}}
}
\keyword{datasets}
